Fix imports for RTD environment
The readthedocs pages were failing to build due to added autodocs that were
trying to import packages that aren't being installed automatically. This
commit adds packages to the requirements.txt file, which RTD uses to build the
environment it compiles the docs in. It also adds a check to see if we're in
the RTD environment to the LS372 Agent to avoid importing OCS.