Improve file creation conditions and file rotation
We don't want to create a file if no frames are being collected, so we rely on
reader.Process() currently blocking before doing any actual writing. We also
want to rotate files for new acquisitions. This will be improved later on when
some more flow control is introduced, but for now we rotate files if the last
frame came in more than 5 seconds ago.