Ensure frame is read before rotating file
This is a quick patch to fix an error state that could occur when the
G3NetworkSender is shutdown that would cause the aggregator to rotate files
every second. While this quick fix solves that problem and should be updated
where used there is still an outstanding problem where the final file will not
be closed until a frame is read (i.e. data collection resumes). This will also
be address shortly.