Make reading frames/making connection not block
We were previously entering a blocking loop on trying to establish a connection
to the G3Reader and on reading frames. This would cause the process to be
impossible to exit. Here we restructure so that no particular method in the
run() loop blocks. This allows the process to be stopped in a timely manner.

We also introduce logic for handling new flow control frames. These are
discarded immediately in read_frames so that they do not break any of the logic
used for splitting acquisitions or creating new files. We also watch for any in
the writer, since we do not want them to end up on disk if they sneak by the
logic somehow.