Make streaming simulator process long lived
This commit makes some adjustments to the way the smurf streaming simulator
process runs, allowing us to send keep alive flow control frames even if we are
not streaming the simulated data frames. This allows us to test some different
scenarios of the streamer states with the timestream aggreagtor.

We also add documentation for running the simulator to the Sphinx docs.