Add heartbeat control task to LS240 agent
This commit adds a single task to the LS240 Agent for testing the disconnection
of an Agent heartbeat. We use the 240 Agent since it is an easy example that
has an accompanying software simulator.

We also adjust the acq method name to conform to naming conventions and update
its documentation.