Add delay in data collection to account for pause
The 372 inputs have a user set parameter called "change pause", which has a
minimum of 3 seconds. This is time after switching channels that the 372 waits
before collecting measurements. If the channel is queried in this time it'll
report the same value repeatedly. This means we end up collecting 3 seconds of
identical, stale, data points. This patch addresses that by tracking when
channels change, querying the pause setting and sleeping for that time.

There's also some other minor cleanup that were simple things I noticed while
linting.