Organize and document cryomech Agent
Added an old page for documentation that I had started working on a while ago.
I also move the agent directory, this is so that sphinx can import the code for
auto-documentation purposes. We should avoid '-' in directory names for this
reason.

I also linted the cryomech agent, there were a lot of PEP08 related spacing
issues.