Adjust the acq method to fit convention and add logging (#26)
This commit is a squash and merge that originally added a heartbeat control task to the 240 Agent. We've since removed that, but still make some useful changes, adjusting the acq method name to conform to naming conventions and updating its documentation. We also add txaio logging and a LOGLEVEL environment variable.