Fix 1 second file creation bug
If the stream was stopped for longer than the 10 second gap timeout, when we
started receiving frames again, we'd generate a new file, then observe that a
frame hadn't come in in longer than the gap, forcing the file closed, then a
new file would be started to write the frames to, since we have frames to write
down now. This would cause a 1 second file to be generated.

It turns out the create_new_file() that this commit removes is superfluous
anyway. A START frame closes the open file, if frames are coming in run() will
still just make a file first thing anyway.