Remove last_meta frame after seeing an END frame
The behavior I was seeing in [1] was caused by the cached last_meta frame.
Clearing it after an END frame is reasonable. This ensures the last_meta frame
is written to each file in a long observation.

We also had to move the Observation frame sending to send every time we start
an observation (i.e. a START frame is sent). Else, we wouldn't get a new meta
frame to write after wiping it.

[1] - https://github.com/simonsobs/socs/pull/46#issuecomment-592707976