Shutdown outside of streaming loop
I got myself stuck in a state where I couldn't shutdown because the streaming
ended before the shutdown flag could get processed, and it was only going to
get processed if in the loop. This was probably rare, and might introduce other
rare states, but I don't think it's a major problem for the simulator.