Move Meinberg interaction to MeinbergSNMP object
Mostly this moves all the direct interaction with the Meinberg to a new object,
MeinbergSNMP. This structure is similar to how several other Agents are
structured. We also introduce a oid_cache, which is used to store the latest
values of each OID and decoded strings for human readability. This cache is
passed to the session.data object for inspection via OCS Clients.