Add operating mode to LogParser for polling for changes
In certain filesystem configurations, checking for inode isn't enough, so we
added checking for modification times on the file. Since this'll hit everytime
even if we didn't need to check this (i.e. "following" the file would work), we
add another environment variable "MODE", to define the operating mode we want
to be in.

This way normal operation like a "tail -f" will be the "follow" mode, and only
open the files once. Otherwise it'll have to be in "poll" mode, which will
reopen the files when changes are detected. More I/O intensive, but these are
only written once a minute, and they're small files.