Drop no:wampy argument and add --cov-config flag
wampy was dropped in ocs v0.8.0, so we no longer need to avoid that wampy
plugin when testing. And to add expand to report coverage in testing of agents
we tell pytest where to look for the coverage file, as it doesn't find it when
running in the container for some reason.