Update run_specified_scan to check free positions
When checking whether the platform has reached the end of a scan,
the Agent will now check whether all of the uploaded points have
been cleared before checking whether the end position is correct.
This will prevent a stop call that comes too early due to scans
over short ranges or completing the line upload when too close to
the stopping position.