Improve functionality in scan_helpers.py generator function
Rename generator function to generate_linear_turnaround. Add keyword
arguments for stop_iter, wait_to_start, step_time, and batch_size,
and implement these throughout the function. Raise errors when
there are values that are not allowed. Change logic to make it
possible to reach the condition where az_flag = 2, and improve logic
for cases where the difference in the endpoints does not divide
evenly by the step size.