Update to the logic within the generator iteration loop
To account for azimuth endpoints whose difference does not divide
neatly by daz, the if/else section now adds the remainder distance
just before the turnaround, ensuring that the difference between
uploaded times is still greater than the minimum 0.05s. Some values
are now rounded to account for divisions that are less neat.

Instead of cycling through the range of stop_iter, implement a while
loop. This allows stop_iter to be set to default float('inf'), while
leaving the option to use a finite stop_iter.

Update the docstring to reflect changes in the parameters.