Handle None values from status feed without setting to 0
When None values are read from the ACU status stream, record them
as float('nan') with the aggregator.

Add new influx feeds for commanded positions on each axis. Only
publish on those feeds when reading a float value from the ACU
status stream.

This update ensures that false zeros are not recorded in g3 files
or shown on grafana.