Kill agent if no SNMP response is received
This can happen in two cases I've seen so far. If there is a mismatch of the
SNMP version used on the M1000 and the one selected in the Agent, in which case
we do want to exit. The other case is if there is a network problem, i.e. the
M1000 is unplugged from the network. In that case we might want to just keep
looping, with the assumption that the network will come back online eventually.