Handle agent misconfiguration and network dropouts
If the SNMP version configured for the Agent and on the M1000 differ, the
snmp.get calls will fail (and look as if the M1000 isn't on the network). So we
add an initial query before entering the acq loop so we can fail early in the
event of misconfiguration.

If a failure occurs during the loop, it's likely a network problem. We keep
looping and record the last conneciton attempt and whether we know the M1000 is
connected or not in session.data. This'll keep the Agent up, and we can alert
on it.