Update _run_specified_scan to abort when stop_and_clear is called
_run_specified_scan now checks whether the number of free upload positions
is greater than 9997, in which case it posts a log warning and ends the Task.
This update is intended to catch cases where the stop_and_clear Task is used
to abort a scan, and prevents _run_specified_scan from continuing to upload
points after stop_and_clear is called.