Remove sleep time before wait for motion
On the simulator the motion can happen within the 5.3 seconds removed
here. This means that when we hit the check for velocity being zero
that waits for 30 seconds for motion to occur, we've already completed
the move and always must wait 30 seconds.

I expect the real ACU takes longer than this 5.3 seconds to start
moving. Removing this should be fine, since instead of waiting outside
the loop we'll be waiting inside the loop for motion. Perhaps the 30
seconds is potentially not long enough?