Adding Ethernet Interface to SCPI_Psui Agent
I added the ethernet_interface.py class so that the SCPI_psu agent can use an ethernet connection instead of being bound to GPIB. I changed around the scpi_psu_driver.py file so that the PsuInterface is no longer direclty inheriting the PrologixInterface class. Instead, there are classes for the prologix and ethernet interfaces that inherit the Prologix/Ethernet Interface. Finally, I changed the agent file so that the parser additionally reads port_number and interface_type (gpib/ethernet) variables from the config. Depending on the interface_type variable, the agent either initializes an ethernet or gpib interface for that PSU.