ACU: go_to uses a state machine loop
Now that we understand the real behaviors of ACU better, it is
possible to write a simpler state-machine to handle the motions.

The rounding and azonly arguments have been removed, as they probably
aren't needed anymore.  The "wait" argument will be removed once
ocs-web stops passing it.

This implementation fixes a couple of bugs: abort works reliably; and
Preset mode isn't set without first setting the target position (which
is a potential problem in some cases, such as when the ACU has just
powered up).