Add num_channels to PsuInterface
This was missing, but needed since it's checked in init(). This is maybe a bit
of a hack, but by setting to 0 we skip setting 'acq_params' and end running
with the default three channel query given by the the default 'channels' on
'monitor_output()'. This is the old behavior before the addition of
'ScpiPsuInterface'.