Add support for direct Ethernet SCPI PSU Agent  (#715)
* Add support for direct ethernet connection to Keithley PSUs, in addition to using the Prologix GPIB-to-ethernet adapter.

* Installing pre-commit and fixing all files with 'pre-commit run --all-files'

* Moving all direct-ethernet-connection changes from prologix_interface.py to drivers.py; adding new ScpiPsuInterface class to implement drivers for SCPI PSUs that can be connected to directly by Ethernet.

* Fixing issues with parsing output from Keithley 2280S-60-3; fixing
problems related to number of channels being different for different
models of PSU.

* Adding '--mode' : 'acq' to the ScpiPsuAgent to enable automatic
acquisition of current and voltage measurements as found in other device
agents.

* fixing issue with psu agent container crashing when attempting to measure when the output channel is off

* Fixing issue related to Setuptools v71 issue #4483; fixing issue where
PSU container would crash when attempting to measure when the output was
disabled.

* Fixing issue with data feed output corrupting influxdb when acq mode used with PSU output disabled.

* Adding PSU agent tasks for getting the output state, voltage
measurements, and current measurements of each channel of the PSU for
use by client scripts.

* Making initial round of changes based on responses to PR #715

* Adding comments to explain the existence of the error code definitions
for the 2280S power supplies.

* Changing output of get_output() to match the other get* functions.

* Fixing issue with call to _initialize_module() in init timing out due to
multiple connection attempts.

* Removing the call to super().__init__(**kwargs) from the
ScpiPsuInterface constructor.

* Start PSU tests in 'init' mode

This was the expected mode before 'acq' became the default. Fixes an issue with
the init test.

* Add expected response to channel enabled query in test

* Add ability to set log level via env var

* Abort monitor process after test

This ends up hanging on the auto-reconnection otherwise. This was a bit
unexpected, as test_mode should break us out before another communication.

* Remove reundant session message

This makes it look like initialize happened twice, since the message included
in the return is also logged in the same place.

* Add test for initializing in 'acq' mode

* Remove commented initialization loop

This was commented because it was causing issues, but we should just drop it
altogether. It's better that 'init' doesn't loop forever until a connection is
made. Rather it should fail and report as failed.

* Add num_channels to PsuInterface

This was missing, but needed since it's checked in init(). This is maybe a bit
of a hack, but by setting to 0 we skip setting 'acq_params' and end running
with the default three channel query given by the the default 'channels' on
'monitor_output()'. This is the old behavior before the addition of
'ScpiPsuInterface'.

---------

Co-authored-by: Daniel Lee <daniel-lee@utexas.edu>
Co-authored-by: Brian Koopman <brian.koopman@yale.edu>