ACU: more care when setting up go_to (#768)
* ACU Agent: make sure clear_faults occurs before moves

This may explain some instances seen on LAT where a go_to leads to
"Irregularity during motion!" error.  The deeper symptom is that the
Preset position appears to not take effect (the "commanded position"
does not update).  So perhaps either an outstanding fault, or a side
effect of clear_faults, prevents the position from updating.

* ACU: when commanding axes in go_to_axes, make sure motions don't clash

This is part (2) of trying to avoid errors where the http connection
is closed or does not reply, when initiating a move.