Enable more verbose codecov reporting
This re-enables comments, patch statuses, the verbose flag in uploads, and does
an ls in the pytest workflow to verify coveage reports are generated properly.

I'll plan to leave this as is for a while until I'm convinced I understand how
codecov works. I think the main problem is the stale PRs don't have base
reports, and rebasing sometimes doesn't actually work to update the reporting.