Correct Task docstrings and fix acq() timeout+typo
Docstrings:
The docstrings for Tasks in this agent were not correctly
formatted. They have been adjusted to accomodate the standards
set forth in the following document:
https://ocs.readthedocs.io/en/main/developer/agent_references/
documentation.html#session-data

Accompanying this reformatting, the session.data object structure
was documented in the acq() docstring.

Timeout:
The timeout for acquiring a lock was set to 100s. This has been
changed to 10s.

Typo:
The self.log.warn String was changed from “Could not start init
because {} is already running” to “Could not start acq because
{} is already running”.