ACU: Support passive axes in go_to (#880)
go_to will now accept az=None and/or el=None, in which case the mode
for the axis will not be changed.  This is an interface change, but
the behavior is backwards compatible with current common usage.  This
is required for star camera surveys with elevation either in Stop or
Preset.