Use new TCPInterface capabilities and move exception handling
Now that PMX is a subclass of TCPInterface, rely on the exception handling
present in the object and mark the session as degraded if needed. I didn't
touch the exception handling when processing individual actions, so this just
covers the "main" process really, not any of the individual tasks, though they
should already be 'robust'.