ACU: generate_scan supports el_mode='stop'
This will cause the el axis to be kept in Stop mode for type 1 and 2
scans.  The default behavior remains the same as it always was -- the
el axis will be in Preset following the initial go_to the start
position.

The az_only argument in generate_scan was not working as advertised,
and does not appear to ever get used, anyway.